moved tensorboard file serialize to base class
